Use consistent output in thermofield function
Previously, the function returned two or four arrays (containing the
chain coefficients) depending on the result of the domain merge.
Moreover, the names of the arguments of the `return`ed tuple changed
from one branch of the `if` block to another, so whoever called the
function had to foresee its outcome.
The new structure of the function is such that the output is always a
dictionary with two keys, `:empty` and `:filled`, which are themselves a
`:frequencies`/`:couplings` tuple just as the other TEDOPA-like function
return. If one of the two environments is empty, the corresponding chain
coefficients will just be zeroes.