Fix evaluation in TFTEDOPA-transformed function
Even when the `indicator` function returns zero, the rest of the product
in the transformed function still gets evaluated, and this is a problem
since the spectral density function is then computed outside its domain
(possibly leading to a DomainError). We solve this by changing the
product

```julia
    indicator(x, set) * f(x)
```

with a conditional

```julia
    iszero(indicator(x, set)) ? zero(x) : f(x)
```

so that `f` is not evaluated at all where it shouldn't be.
Since `set` might not be a simple `[a, b]` list, the condition
`iszero(indicator(x, set))` is not always reducible to something like
`a <= x <= b` so this syntax, even if it is somewhat complicated, works
with more generic `set`s.